Fix YAML annotation
Don't set the method option "tag" so examples/internal/clients/unannotatedecho/api/swagger.yaml
generates a examples/internal/clients/unannotatedecho/api_unannotated_echo_service.go that works with existing tests.
